+++ service/handlers.go
@@ -12,13 +12,14 @@ import (
 	"strings"
 
 	"github.com/argoproj-labs/argo-cloudops/internal/env"
+	"github.com/argoproj-labs/argo-cloudops/service/internal/credentials"
 	"github.com/argoproj-labs/argo-cloudops/service/internal/workflow"
+
 	"github.com/aws/aws-sdk-go/aws/arn"
 	"github.com/distribution/distribution/reference"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/gorilla/mux"
-	vault "github.com/hashicorp/vault/api"
 	"gopkg.in/yaml.v2"
 )
 
@@ -68,62 +69,15 @@ var isStringAlphaNumeric = regexp.MustCompile(`^[a-zA-Z0-9_]*$`).MatchString
 // Vault does not allow for dashes
 var isStringAlphaNumericUnderscore = regexp.MustCompile(`^([a-zA-Z])[a-zA-Z0-9_]*$`).MatchString
 
-// Authorization represents a user's authorization token.
-type Authorization struct {
-	Provider string
-	Key      string
-	Secret   string
-}
-
-// Authorization function for token requests.
-// This is separate from admin functions which use the admin env var
-func newAuthorization(authorizationHeader string) (*Authorization, error) {
-	var a Authorization
-	auth := strings.SplitN(authorizationHeader, ":", 3)
-	for _, i := range auth {
-		if i == "" {
-			return nil, fmt.Errorf("invalid authorization header provided")
-		}
-	}
-	if len(auth) < 3 {
-		return nil, fmt.Errorf("invalid authorization header provided")
-	}
-	a.Provider = auth[0]
-	a.Key = auth[1]
-	a.Secret = auth[2]
-	return &a, nil
-}
-
-// Returns true, if the user is an admin.
-func (a Authorization) isAdmin() bool {
-	return a.Key == "admin"
-}
-
-// Returns true, if the user is an authorized admin
-func (a Authorization) authorizedAdmin(adminSecret string) bool {
-	return a.isAdmin() && a.Secret == adminSecret
-}
-
 // HTTP handler
 type handler struct {
 	logger                 log.Logger
-	newCredentialsProvider func(a Authorization) (credentialsProvider, error)
+	newCredentialsProvider func(a credentials.Authorization) (credentials.Provider, error)
 	argo                   workflow.Workflow
 	argoCtx                context.Context
 	config                 *Config
 	gitClient              gitClient
 	env                    env.EnvVars
-}
-
-// Returns a new vaultCredentialsProvider
-func newVaultProvider(svc *vault.Client) func(a Authorization) (credentialsProvider, error) {
-	return func(a Authorization) (credentialsProvider, error) {
-		return &vaultCredentialsProvider{
-			VaultSvc: svc,
-			RoleID:   a.Key,
-			SecretID: a.Secret,
-		}, nil
-	}
 }
 
 // Validates workflow parameters
@@ -239,7 +193,7 @@ func (h handler) createWorkflowFromGit(w http.ResponseWriter, r *http.Request) {
 	}
 
 	ah := r.Header.Get("Authorization")
-	a, err := newAuthorization(ah)
+	a, err := credentials.NewAuthorization(ah)
 	if err != nil {
 		h.errorResponse(w, "error authorizing", http.StatusUnauthorized, err)
 		return
@@ -284,7 +238,7 @@ func (h handler) createWorkflow(w http.ResponseWriter, r *http.Request) {
 	log.With(l, "project", cwr.ProjectName, "target", cwr.TargetName, "framework", cwr.Framework, "type", cwr.Type, "workflow-template", cwr.WorkflowTemplateName)
 
 	ah := r.Header.Get("Authorization")
-	a, err := newAuthorization(ah)
+	a, err := credentials.NewAuthorization(ah)
 	if err != nil {
 		h.errorResponse(w, "error authorizing", http.StatusUnauthorized, err)
 		return
@@ -295,7 +249,7 @@ func (h handler) createWorkflow(w http.ResponseWriter, r *http.Request) {
 }
 
 // Creates a workflow
-func (h handler) createWorkflowFromRequest(ctx context.Context, w http.ResponseWriter, a *Authorization, cwr createWorkflowRequest, l log.Logger) {
+func (h handler) createWorkflowFromRequest(ctx context.Context, w http.ResponseWriter, a *credentials.Authorization, cwr createWorkflowRequest, l log.Logger) {
 	level.Debug(l).Log("message", "validating workflow parameters")
 	if err := h.validateWorkflowParameters(cwr.Parameters); err != nil {
 		level.Error(l).Log("message", "error in parameters", "error", err)
@@ -384,14 +338,14 @@ func (h handler) createWorkflowFromRequest(ctx context.Context, w http.ResponseW
 	}
 
 	level.Debug(l).Log("message", "getting credentials provider token")
-	credentialsToken, err := cp.getToken()
+	credentialsToken, err := cp.GetToken()
 	if err != nil {
 		level.Error(l).Log("message", "error getting credentials provider token", "error", err)
 		h.errorResponse(w, "error getting credentials provider token", http.StatusInternalServerError, err)
 		return
 	}
 
-	projectExists, err := cp.projectExists(cwr.ProjectName)
+	projectExists, err := cp.ProjectExists(cwr.ProjectName)
 	if err != nil {
 		level.Error(l).Log("message", "error checking project", "error", err)
 		h.errorResponse(w, "error checking project", http.StatusInternalServerError, err)
@@ -477,7 +431,7 @@ func (h handler) getTarget(w http.ResponseWriter, r *http.Request) {
 
 	level.Debug(l).Log("message", "authorizing get target permissions")
 	ah := r.Header.Get("Authorization")
-	a, err := newAuthorization(ah) // todo add validation
+	a, err := credentials.NewAuthorization(ah) // todo add validation
 	if err != nil {
 		level.Error(l).Log("message", "error authorizing using Authorization header", "error", err)
 		h.errorResponse(w, "error authorizing using Authorization header", http.StatusUnauthorized, err)
@@ -485,7 +439,7 @@ func (h handler) getTarget(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "validating authorized admin")
-	if !a.authorizedAdmin(h.env.AdminSecret) {
+	if !a.AuthorizedAdmin(h.env.AdminSecret) {
 		level.Error(l).Log("message", "must be an authorized admin")
 		h.errorResponse(w, "must be an authorized admin", http.StatusUnauthorized, nil)
 		return
@@ -500,7 +454,7 @@ func (h handler) getTarget(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "getting target information")
-	targetInfo, err := cp.getTarget(projectName, targetName)
+	targetInfo, err := cp.GetTarget(projectName, targetName)
 	if err != nil {
 		level.Error(l).Log("message", "error retrieving target information", "error", err)
 		h.errorResponse(w, "error retrieving target information", http.StatusBadRequest, err)
@@ -584,7 +538,7 @@ func (h handler) createProject(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var capp createProjectRequest
+	var capp credentials.CreateProjectRequest
 	err = json.Unmarshal(reqBody, &capp)
 	if err != nil {
 		level.Error(l).Log("message", "error parsing json", "error", err)
@@ -596,7 +550,7 @@ func (h handler) createProject(w http.ResponseWriter, r *http.Request) {
 
 	level.Debug(l).Log("message", "authorizing project creation")
 	ah := r.Header.Get("Authorization")
-	a, err := newAuthorization(ah) // todo add validation
+	a, err := credentials.NewAuthorization(ah) // todo add validation
 	if err != nil {
 		level.Error(l).Log("message", "error authorizing using Authorization header", "error", err)
 		h.errorResponse(w, "error authorizing using Authorization header", http.StatusUnauthorized, err)
@@ -604,7 +558,7 @@ func (h handler) createProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "validating authorized admin")
-	if !a.authorizedAdmin(h.env.AdminSecret) {
+	if !a.AuthorizedAdmin(h.env.AdminSecret) {
 		level.Error(l).Log("message", "must be an authorized admin")
 		h.errorResponse(w, "must be an authorized admin", http.StatusUnauthorized, nil)
 		return
@@ -624,7 +578,7 @@ func (h handler) createProject(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	projectExists, err := cp.projectExists(capp.Name)
+	projectExists, err := cp.ProjectExists(capp.Name)
 	if err != nil {
 		level.Error(l).Log("message", "error checking project", "error", err)
 		h.errorResponse(w, "error checking project", http.StatusInternalServerError, err)
@@ -638,7 +592,7 @@ func (h handler) createProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "creating project")
-	role, secret, err := cp.createProject(capp.Name)
+	role, secret, err := cp.CreateProject(capp.Name)
 	if err != nil {
 		level.Error(l).Log("message", "error creating project", "error", err)
 		h.errorResponse(w, "error creating project", http.StatusInternalServerError, err)
@@ -665,7 +619,7 @@ func (h handler) getProject(w http.ResponseWriter, r *http.Request) {
 
 	level.Debug(l).Log("message", "authorizing get project")
 	ah := r.Header.Get("Authorization")
-	a, err := newAuthorization(ah) // todo add validation
+	a, err := credentials.NewAuthorization(ah) // todo add validation
 	if err != nil {
 		level.Error(l).Log("message", "error authorizing using Authorization token", "error", err)
 		h.errorResponse(w, "error authorizing using Authorization header", http.StatusUnauthorized, err)
@@ -673,7 +627,7 @@ func (h handler) getProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "validating authorized admin")
-	if !a.authorizedAdmin(h.env.AdminSecret) {
+	if !a.AuthorizedAdmin(h.env.AdminSecret) {
 		level.Error(l).Log("message", "must be an authorized admin")
 		h.errorResponse(w, "must be an authorized admin", http.StatusUnauthorized, nil)
 		return
@@ -688,7 +642,7 @@ func (h handler) getProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "getting project")
-	jsonResult, err := cp.getProject(projectName)
+	jsonResult, err := cp.GetProject(projectName)
 	if err != nil {
 		level.Error(l).Log("message", "error retrieving project", "error", err)
 		h.errorResponse(w, "error retrieving project", http.StatusNotFound, err)
@@ -706,7 +660,7 @@ func (h handler) deleteProject(w http.ResponseWriter, r *http.Request) {
 
 	level.Debug(l).Log("message", "authorizing delete project")
 	ah := r.Header.Get("Authorization")
-	a, err := newAuthorization(ah) // todo add validation
+	a, err := credentials.NewAuthorization(ah) // todo add validation
 	if err != nil {
 		level.Error(l).Log("message", "error authorizing using Authorization header", "error", err)
 		h.errorResponse(w, "error authorizing using Authorization header", http.StatusUnauthorized, err)
@@ -714,7 +668,7 @@ func (h handler) deleteProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "validating authorized admin")
-	if !a.authorizedAdmin(h.env.AdminSecret) {
+	if !a.AuthorizedAdmin(h.env.AdminSecret) {
 		level.Error(l).Log("message", "must be an authorized admin")
 		h.errorResponse(w, "must be an authorized admin", http.StatusUnauthorized, nil)
 		return
@@ -729,7 +683,7 @@ func (h handler) deleteProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "checking if project exists")
-	projectExists, err := cp.projectExists(projectName)
+	projectExists, err := cp.ProjectExists(projectName)
 	if err != nil {
 		level.Error(l).Log("message", "error checking project", "error", err)
 		h.errorResponse(w, "error checking project", http.StatusInternalServerError, err)
@@ -742,7 +696,7 @@ func (h handler) deleteProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "getting all targets in project")
-	targets, err := cp.listTargets(projectName)
+	targets, err := cp.ListTargets(projectName)
 	if err != nil {
 		level.Error(l).Log("message", "error getting all targets", "error", err)
 		h.errorResponse(w, "error getting all targets", http.StatusInternalServerError, err)
@@ -756,7 +710,7 @@ func (h handler) deleteProject(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "deleting project")
-	err = cp.deleteProject(projectName)
+	err = cp.DeleteProject(projectName)
 	if err != nil {
 		level.Error(l).Log("message", "error deleting project", "error", err)
 		h.errorResponse(w, "error deleting project", http.StatusBadRequest, err)
@@ -780,7 +734,7 @@ func (h handler) createTarget(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var ctr createTargetRequest
+	var ctr credentials.CreateTargetRequest
 	err = json.Unmarshal(reqBody, &ctr)
 	if err != nil {
 		level.Error(l).Log("message", "error parsing request body to target data", "error", err)
@@ -791,7 +745,7 @@ func (h handler) createTarget(w http.ResponseWriter, r *http.Request) {
 	l = log.With(l, "target", ctr.Name)
 
 	level.Debug(l).Log("message", "authorizing target creation")
-	a, err := newAuthorization(ah)
+	a, err := credentials.NewAuthorization(ah)
 	if err != nil {
 		level.Error(l).Log("message", "error authorizing using Authorization header", "error", err)
 		h.errorResponse(w, "error authorizing using Authorization header", http.StatusUnauthorized, err)
@@ -799,7 +753,8 @@ func (h handler) createTarget(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "validating authorized admin")
-	if !a.authorizedAdmin(h.env.AdminSecret) {
+
+	if !a.AuthorizedAdmin(h.env.AdminSecret) {
 		level.Error(l).Log("message", "must be an authorized admin")
 		h.errorResponse(w, "must be an authorized admin", http.StatusUnauthorized, nil)
 		return
@@ -846,7 +801,7 @@ func (h handler) createTarget(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	targetExists, _ := cp.targetExists(ctr.Name)
+	targetExists, _ := cp.TargetExists(ctr.Name)
 	// TODO: handle error when implemented
 	if targetExists {
 		level.Error(l).Log("message", "target name must not already exist")
@@ -855,7 +810,7 @@ func (h handler) createTarget(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "creating target")
-	err = cp.createTarget(projectName, ctr)
+	err = cp.CreateTarget(projectName, ctr)
 	if err != nil {
 		level.Error(l).Log("message", "error creating target", "error", err)
 		h.errorResponse(w, "error creating target", http.StatusInternalServerError, err)
@@ -875,7 +830,7 @@ func (h handler) deleteTarget(w http.ResponseWriter, r *http.Request) {
 
 	level.Debug(l).Log("message", "authorizing delete target permissions")
 	ah := r.Header.Get("Authorization")
-	a, err := newAuthorization(ah) // todo add validation
+	a, err := credentials.NewAuthorization(ah) // todo add validation
 	if err != nil {
 		level.Error(l).Log("message", "error authorizing using Authorization header", "error", err)
 		h.errorResponse(w, "error authorizing using Authorization header", http.StatusUnauthorized, err)
@@ -883,7 +838,8 @@ func (h handler) deleteTarget(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "validating authorized admin")
-	if !a.authorizedAdmin(h.env.AdminSecret) {
+
+	if !a.AuthorizedAdmin(h.env.AdminSecret) {
 		level.Error(l).Log("message", "must be an authorized admin")
 		h.errorResponse(w, "must be an authorized admin", http.StatusUnauthorized, nil)
 		return
@@ -898,7 +854,7 @@ func (h handler) deleteTarget(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "deleting target")
-	err = cp.deleteTarget(projectName, targetName)
+	err = cp.DeleteTarget(projectName, targetName)
 	if err != nil {
 		level.Error(l).Log("message", "error deleting target", "error", err)
 		h.errorResponse(w, "error deleting target", http.StatusBadRequest, err)
@@ -915,7 +871,7 @@ func (h handler) listTargets(w http.ResponseWriter, r *http.Request) {
 	l := h.requestLogger(r, "op", "list-targets", "project", projectName)
 
 	level.Debug(l).Log("message", "authorizing target list retrieval")
-	a, err := newAuthorization(ah)
+	a, err := credentials.NewAuthorization(ah)
 	if err != nil {
 		level.Error(l).Log("message", "error authorizing using Authorization header", "error", err)
 		h.errorResponse(w, "error authorizing using Authorization header", http.StatusUnauthorized, err)
@@ -923,7 +879,8 @@ func (h handler) listTargets(w http.ResponseWriter, r *http.Request) {
 	}
 
 	level.Debug(l).Log("message", "validating authorized admin")
-	if !a.authorizedAdmin(h.env.AdminSecret) {
+
+	if !a.AuthorizedAdmin(h.env.AdminSecret) {
 		level.Error(l).Log("message", "must be an authorized admin")
 		h.errorResponse(w, "must be an authorized admin", http.StatusUnauthorized, err)
 		return
@@ -937,7 +894,7 @@ func (h handler) listTargets(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	targets, err := cp.listTargets(projectName)
+	targets, err := cp.ListTargets(projectName)
 	if err != nil {
 		level.Error(l).Log("message", "error listing targets", "error", err)
 		h.errorResponse(w, "error listing targets", http.StatusInternalServerError, err)