+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.1.0 (2021-06-08)
+### Added
+* Tests for vault credential provider
+* Vault service health check
+
+### Changed
+* Update credentials provider to be internal package
+* Environmental variable handling
+
 ## v0.0.4 (2021-06-03)
 ### Fixed
 * Passing Argo context to Argo Workflow calls